+++ common/engine/khg_phy/include/khg_phy/contact.h
@@ -19,7 +19,7 @@ typedef struct phy_contact {
   phy_vector2 anchor_a;
   phy_vector2 anchor_b;
   float separation;
-  unsigned long id;
+  unsigned long long id;
   bool is_persisted;
   bool remove_invoked;
   phy_contact_solver_info solver_info;
@@ -45,7 +45,7 @@ typedef struct phy_contact_event {
   phy_vector2 position;
   phy_vector2 normal_impulse;
   phy_vector2 friction_impulse;
-  unsigned long id;
+  unsigned long long id;
 } phy_contact_event;
 
 typedef void (*phy_contact_listener_callback)(struct phy_space *space, phy_contact_event event, void *user_arg);
@@ -58,13 +58,13 @@ typedef struct phy_contact_listener {
 
 static const phy_contact_solver_info phy_contact_solver_info_zero = { 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0 };
 
-static inline unsigned long phy_persistent_contact_pair_key(phy_shape *a, phy_shape *b) {
+static inline unsigned long long phy_persistent_contact_pair_key(phy_shape *a, phy_shape *b) {
   unsigned int fpa = phy_u32_hash(a->id);
   unsigned int fpb = phy_u32_hash(b->id);
   return phy_u32_pair(fpa, fpb);
 }
 
 bool phy_persistent_contact_pair_penetrating(phy_persistent_contact_pair *pcp);
-unsigned long phy_persistent_contact_pair_hash(void *item);
+unsigned long long phy_persistent_contact_pair_hash(void *item);
 void phy_persistent_contact_pair_remove(struct phy_space *space, phy_persistent_contact_pair *pcp);
